feat: Use dcurl shared library from JAR file
DO NOT MERGE

The dcurl submodule is updated to support generating dcurl JAR file
and the Makefile can install the JAR file.
The priority to load the shared library:
1. Search from the java.library.path and load it. (Deprecated)
2. Load from the installed JAR file.
If the dcurl library is unable to use, the IRI prints out the
information that there is no external library for PearlDiver.